Split runner into 2 classes in readiness for runDirectoryOfApsimFiles
