+++ Models/Models.csproj
@@ -230,6 +230,7 @@
     <Compile Include="Core\Run\CommenceArgs.cs" />
     <Compile Include="Core\Run\GenerateApsimXFiles.cs" />
     <Compile Include="Core\Run\IReplacement.cs" />
+    <Compile Include="Core\Run\JobCollection.cs" />
     <Compile Include="Core\Run\ModelReplacement.cs" />
     <Compile Include="Core\Run\PropertyReplacement.cs" />
     <Compile Include="Core\Run\Runner.cs" />
@@ -255,9 +256,7 @@
     <Compile Include="Core\Links.cs" />
     <Compile Include="Core\ModelTypes.cs" />
     <Compile Include="Core\Runners\RunDirectoryOfApsimFiles.cs" />
-    <Compile Include="Core\Runners\RunSimulation.cs" />
     <Compile Include="Core\Runners\RunExternal.cs" />
-    <Compile Include="Core\Runners\RunOrganiser.cs" />
     <Compile Include="Core\Scope.cs" />
     <Compile Include="Core\VariableMethod.cs" />
     <Compile Include="Functions\AccumulateAtEvent.cs" />
